requirements.txt
requests
docx2txt
pdfminer.six
spacy
bcrypt
langchain
langchain-community
faiss-cpu
pypdf
sentence-transformers
agent
google-api-python-client 
google-auth 
google-auth-httplib2 
google-auth-oauthlib